[Feature] Display Movie Detail in Tablet or Two-Pane Mode

+++ app/src/main/java/au/com/appetiser/isearch/database/MovieDao.kt
@@ -17,6 +17,6 @@ interface MovieDao {
     fun getAllMovies(): LiveData<List<Movie>>
 
     @Query("SELECT * FROM movie_table WHERE trackId = :trackId")
-    fun getMovieById(trackId: Int): Movie
+    fun getMovieById(trackId: Long): Movie
 
 }
